FE (themes) - Relocated themes into globals file
