+++ frontend/src/App.js
@@ -4,6 +4,9 @@ import { connect } from 'react-redux';
 import { scroller } from 'react-scroll';
 import styled, { createGlobalStyle, ThemeProvider } from 'styled-components';
 
+// themes
+import {dayTheme, nightTheme} from './globals/globals';
+
 // components
 import {
   Header,
@@ -55,72 +58,8 @@ const GlobalStyle = createGlobalStyle`
 	}
 `;
 
-const dayTheme = {
-  appBgColor:  '#54bdff',
-  authBgColor: 'gray',
-  authColor: 'white',
-  authLinkRegColor: 'white',
-  authLinkRegColorHov: 'black',
-  authLoginColor: 'white',
-  authLoginColorHov: 'black',
-  catNameColor: 'black',
-  catDiscussionCountColor: 'black',
-  catBgColorHov: 'rgba(255, 255, 255, 0.6)',
-  catTimestampColor: 'black',
-  catTitleColor:'black',
-  catNameDateColor:'black',
-  discussionUsernameColor: 'black',
-  discussionByCatWrapperBgColor: '#e8e3e0',
-  discussionByCatWrapperBgColorHov: 'rgba(255, 255, 255, 0.195)',
-  discussionByCatWrapperBgShdw:'2px 3px 2px 2px #610b07',
-  discussionByCatTitleColor:'black',
-  discussionByCatTitleBgColorHov:'rgba(255, 255, 255, 0.13)',
-  discussionByCatTitleColorHov:'white',
-  discussionByCatCategoryColor: 'black',
-  discussionByCatCategoryBgColorHov: 'rgba(255, 255, 255, 0.13)',
-  discussionByCatCategoryColorHov: 'white',
-  discussionByCatNameDateColor: 'black',
-  discussionByCatNameDateBgColorHov: 'rgba(255, 255, 255, 0.13)',
-  discussionByCatNameDateColorHov: 'white',
-  skyColor: '#37d8e6',
-  symposiumProfileBgColor: '#ffdd00',
-  symposiumBorderColor: '#f1c40f',
-};
 
-const nightTheme = {
-  appBgColor: 'pink',
-  authBgColor: 'black',
-  authColor: 'gray',
-  authLinkRegColor: 'black',
-  authLinkRegColorHov: 'white',
-  authLoginColor: 'black',
-  authLoginColorHov: 'white',
-  catNameColor: 'red',
-  catDiscussionCountColor: 'red',
-  catBgColorHov: 'rgba(100, 200, 200, 0.9)',
-  catTimestampColor: 'red',
-  catTitleColor: 'red',
-  catNameDateColor:'red',
-  discussionUsernameColor: 'white',
-  discussionByCatWrapperBgColor: 'red',
-  discussionByCatWrapperBgColorHov: 'rgba(100, 200, 255, 0.33)',
-  discussionByCatWrapperBgShdw:'2px 3px 2px 2px pink',
-  discussionByCatTitleColor:'white',
-  discussionByCatTitleBgColorHov:'rgba(100, 200, 255, 0.33)',
-  discussionByCatTitleColorHov:'black',
-  discussionByCatCategoryColor: 'white',
-  discussionByCatCategoryBgColorHov: 'rgba(100, 200, 255, 0.33)',
-  discussionByCatCategoryColorHov: 'black',
-  discussionByCatNameDateColor: 'white',
-  discussionByCatNameDateBgColorHov: 'rgba(100, 200, 255, 0.33)',
-  discussionByCatNameDateColorHov: 'black',
-  skyColor: '#2c3e50',
-  symposiumProfileBgColor: '#bdc3c7',
-  symposiumBorderColor: '#eaeff2',
-}
-
-class App extends Component {
-  
+class App extends Component {  
   constructor(props) {
     super(props);
 
@@ -203,7 +142,7 @@ class App extends Component {
           <Header history={history} toggleSearch = { this.toggleSearch } />
           <button
           onClick={() => this.handleClick()}>
-          Dark Theme</button>
+          Switch Themes</button>
           <Switch>
             <Route path='/register' component={RegisterView} />
             <Route path='/request-reset-pw' component={RequestResetPWForm} />
